+++ problem4j-spring-web/src/test/kotlin/io/github/problem4j/spring/web/resolver/ProblemResolverExtensionsTest.kt
@@ -16,7 +16,6 @@
 
 package io.github.problem4j.spring.web.resolver
 
-import io.github.problem4j.core.Problem
 import io.github.problem4j.core.ProblemContext
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
@@ -28,7 +27,10 @@ class ProblemResolverExtensionsTest {
   @Test
   fun givenExceptionSubtype_whenResolvingWithFunctionalResolver_thenTypedExceptionIsPassed() {
     val resolver =
-        problemResolver<IllegalStateException> { _, ex, _, _ -> Problem.of(500, ex.message) }
+        problemResolver<IllegalStateException> { _, ex ->
+          status(500)
+          detail(ex.message)
+        }
 
     assertThat(resolver.exceptionClass).isEqualTo(IllegalStateException::class.java)